Renaming webpack.config in webpack.dev to match scripts
