+++ README.md
@@ -138,7 +138,7 @@ The default webpack config shipped with scriptoni should be fine in most cases. 
 If this is the case, you can override the default config by passing an additional `--webpackConfig` argument, followed by the file path containing your override function.
 
 Let's say, for example, you want to change the output library.
-You can provide a `webpack.config.js` file in the root directory of your project, with the following content:
+You can provide a `webpack.dev.js` file in the root directory of your project, with the following content:
 
 ```js
 module.exports = (defaultConfig, { config, paths, NODE_ENV, target }) => ({
@@ -156,7 +156,7 @@ As you can see, your function will receive the default webpack config as first a
 - `NODE_ENV`: 'development' or 'production'
 - `target`: one of `dev`, `build`, `dev-ts` or `build-ts`
 
-As a last step, you can change the `start` and `build` scripts in your `package.json` file by adding `--webpackConfig ./webpack.config.js` and the end of both commands:
+As a last step, you can change the `start` and `build` scripts in your `package.json` file by adding `--webpackConfig ./webpack.dev.js` and the end of both commands:
 
 ```json
 "start": "UV_THREADPOOL_SIZE=20 scriptoni web-dev -c ./config --webpackConfig ./webpack-config.js",
